refactor: Move files logic into its own module
